+++ service/controller/resource/monitoring/remotewriteingressauth/client.go
@@ -1,4 +1,4 @@
-package remotewriteapiendpointsecret
+package remotewriteingressauth
 
 import (
 	"context"