+++ CHANGELOG.md
@@ -7,10 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.6.4] - 2022-10-17
+
+### Changed
+
+- Move shield route so alerts for shield don't go to opsgenie at all, only to their slack.
+
 ## [4.6.3] - 2022-10-17
 
 ### Changed
 
+- Change remotewrite endpoint secrets namespace to clusterID ns.
 - Add `.svc` suffix to the alertmanager target to make PMO work behind a corporate proxy.
 - Upgrade to go 1.19
 - Bump prometheus-operator to v0.54.0
@@ -1621,7 +1628,8 @@ This release was created on release-v3.5.x branch to fix release 3.6.0 see PR#99
 - First release.
 
 
-[Unreleased]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.3...HEAD
+[Unreleased]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.4...HEAD
+[4.6.4]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.3...v4.6.4
 [4.6.3]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.2...v4.6.3
 [4.6.2]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.1...v4.6.2
 [4.6.1]: https://github.com/giantswarm/prometheus-meta-operator/compare/v4.6.0...v4.6.1